@@ -150,7 +150,6 @@ class ShowMoneyFrame extends StatelessWidget {
       return Row(
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: [
-          
           Text(
             getTranslated(context, typeName)!,
             style: TextStyle(fontSize: 22.sp),
@@ -184,18 +183,16 @@ class ShowMoneyFrame extends StatelessWidget {
       child: Padding(
         padding: EdgeInsets.symmetric(horizontal: 20.w, vertical: 18.5.h),
         child: Column(
-          
           children: [
-            
             rowFrame(this.type, typeValue),
             SizedBox(
               height: 12.5.h,
             ),
             rowFrame('Balance', this.balance),
-                        SizedBox(
+            SizedBox(
               height: 12.5.h,
             ),
-           rowFrame('Total', typeValue + balance),
+            rowFrame('Total', typeValue + balance),
             SizedBox(
               height: 12.5.h,
             ),
