@@ -15,7 +15,6 @@ class SelectLanguage extends StatefulWidget {
 }
 
 class _SelectLanguageState extends State<SelectLanguage> {
-
   @override
   Widget build(BuildContext context) {
     List<Language> languageList = Language.languageList;
@@ -35,73 +34,64 @@ class _SelectLanguageState extends State<SelectLanguage> {
                   onPressed: () => Navigator.pop(context),
                 ),
               )
-            ]
-        ),
+            ]),
         body: ChangeNotifierProvider<OnLanguageSelected>(
-          create: (context) => OnLanguageSelected(),
-          builder: (context, widget) => Container(
-            color : white,
-              child: 
-              
-              
-                      ListView.builder(
-              itemCount: languageList.length,
-              itemBuilder: (context, int) {
-                return GestureDetector(
-                  behavior: HitTestBehavior.translucent,
-                  onTap: () {
-                    Locale _locale = sharedPrefs.setLocale(languageList[int].languageCode);
-                    MyApp.setLocale(context, _locale);
-                    context
-                        .read<OnLanguageSelected>()
-                        .onSelect(languageList[int].languageCode);
-                  },
-                  child: Column(
-                    children: [
-                      Padding(
-                        padding: EdgeInsets.symmetric(
-                            vertical: 5.h, horizontal: 23.w),
-                        child: Row(
-                          children: [
-                            Text(
-                              languageList[int].flag,
-                              style: TextStyle(fontSize: 45.sp),
-                            ),
-                            SizedBox(
-                              width: 35.w,
-                            ),
-                            Text(languageList[int].name,
-                                style: TextStyle(
-                                  fontSize: 20.sp,
-                                )),
-                            Spacer(),
-                            context.watch<OnLanguageSelected>().languageCode ==
-                                    languageList[int].languageCode
-                                ? Icon(Icons.check_circle,
-                                    size: 25.sp, color: blue3)
-                                : SizedBox(),
-                            SizedBox(width: 15.w)
-                          ],
-                        ),
-                      ),
-                      Divider(
-                        indent: 90.w,
-                        height: 0,
-                        thickness: 0.25.h,
-                        color: grey,
-                      ),
-                    ],
-                  ),
-                );
-              }),
-          
-          )
-          
-          
-  
-        
-        
-        
-        ));
+            create: (context) => OnLanguageSelected(),
+            builder: (context, widget) => Container(
+                  color: white,
+                  child: ListView.builder(
+                      itemCount: languageList.length,
+                      itemBuilder: (context, int) {
+                        return GestureDetector(
+                          behavior: HitTestBehavior.translucent,
+                          onTap: () {
+                            Locale _locale = sharedPrefs
+                                .setLocale(languageList[int].languageCode);
+                            MyApp.setLocale(context, _locale);
+                            context
+                                .read<OnLanguageSelected>()
+                                .onSelect(languageList[int].languageCode);
+                          },
+                          child: Column(
+                            children: [
+                              Padding(
+                                padding: EdgeInsets.symmetric(
+                                    vertical: 5.h, horizontal: 23.w),
+                                child: Row(
+                                  children: [
+                                    Text(
+                                      languageList[int].flag,
+                                      style: TextStyle(fontSize: 45.sp),
+                                    ),
+                                    SizedBox(
+                                      width: 35.w,
+                                    ),
+                                    Text(languageList[int].name,
+                                        style: TextStyle(
+                                          fontSize: 20.sp,
+                                        )),
+                                    Spacer(),
+                                    context
+                                                .watch<OnLanguageSelected>()
+                                                .languageCode ==
+                                            languageList[int].languageCode
+                                        ? Icon(Icons.check_circle,
+                                            size: 25.sp, color: blue3)
+                                        : SizedBox(),
+                                    SizedBox(width: 15.w)
+                                  ],
+                                ),
+                              ),
+                              Divider(
+                                indent: 90.w,
+                                height: 0,
+                                thickness: 0.25.h,
+                                color: grey,
+                              ),
+                            ],
+                          ),
+                        );
+                      }),
+                )));
   }
 }
